backend/requirements.txt
Django
djangorestframework
google-generativeai
python-dotenv
django-cors-headers
psycopg2-binary
dj-database-url
